% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotivol.R
\name{compOptionPrice}
\alias{compOptionPrice}
\title{Compute the undiscounted price of an option with Black76 model.}
\usage{
compOptionPrice(option_type, future, strike, vol, time_to_exp)
}
\arguments{
\item{option_type}{type of the option, either a call or a put as 'C' or 'P'.}

\item{future}{price of the future contract.}

\item{strike}{strike of the option.}

\item{vol}{implied volatility.}

\item{time_to_expiry}{remaining life of the option, in fraction of a year.}
}
\value{
The undiscounted price of a \code{option_type} of strike \code{strike} and option life of \code{time_to_expiry}.
}
\description{
/code{compOptionPrice} computes the price of an option on futures. The price is not discounted.
}
\examples{
df <- data.frame(strike = c(50, 20), # the option strike - in $
                 type = c("C", "P"), # either “c” for call option or “p” for a put option
                 futurePrice = c(48.03, 48.03), # the price of the underlying future - in $
                 vol = c(0.08, 0.30) # the implied volatility of the underlying future
                 time_to_expiry = c(0.1423, 0.1423))

compOptionPrice(df)
}
